requirements.txt
google-adk>=0.5.0
google-genai>=1.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
flask>=3.0.0
gunicorn>=22.0.0
python-dateutil>=2.9.0
